test(IDX-3144): inline execution system-tests
