+++ rs/tests/execution/common/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "execution-system-tests-common"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description.workspace = true
+documentation.workspace = true
+
+[dependencies]
+ic-registry-subnet-type = { path = "../../../registry/subnet_type" }
+ic-system-test-driver = { path = "../../driver" }
+
+[lib]
+name = "execution_system_tests_common"
+path = "execution_system_tests_common.rs"
