+++ rs/tests/execution/common/execution_system_tests_common.rs
@@ -1,7 +1,3 @@
-pub mod call_on_cleanup;
-pub mod inter_canister_queries;
-pub mod system_api_security_test;
-
 use ic_registry_subnet_type::SubnetType;
 use ic_system_test_driver::driver::{
     ic::{InternetComputer, Subnet},
