+++ rs/tests/execution/cycles_restrictions_test.rs
@@ -1,9 +1,16 @@
 use anyhow::Result;
-
+use execution_system_tests_common::config_system_verified_application_subnets;
+use ic_agent::{agent::RejectCode, export::Principal, AgentError};
 use ic_system_test_driver::driver::group::SystemTestGroup;
+use ic_system_test_driver::driver::test_env::TestEnv;
+use ic_system_test_driver::driver::test_env_api::{GetFirstHealthyNodeSnapshot, HasPublicApiUrl};
 use ic_system_test_driver::systest;
-use ic_tests::execution::canister_lifecycle::controller_and_controllee_on_different_subnets;
-use ic_tests::execution::config_system_verified_application_subnets;
+use ic_system_test_driver::util::{
+    assert_http_submit_fails, assert_reject, block_on, set_controller, UniversalCanister,
+};
+use ic_universal_canister::{management, wasm, UNIVERSAL_CANISTER_WASM};
+use ic_utils::call::AsyncCall;
+use ic_utils::interfaces::ManagementCanister;
 
 fn main() -> Result<()> {
     SystemTestGroup::new()
@@ -12,4 +19,83 @@ fn main() -> Result<()> {
         .execute_from_args()?;
 
     Ok(())
+}
+
+/// A test to ensure that controller and controllee canisters can exist on
+/// different subnets and they can still control each other.
+pub fn controller_and_controllee_on_different_subnets(env: TestEnv) {
+    let logger = env.logger();
+    let ver_app_node = env.get_first_healthy_verified_application_node_snapshot();
+    let app_node = env.get_first_healthy_application_node_snapshot();
+    let combinations = vec![
+        (ver_app_node.clone(), app_node.clone()),
+        (app_node, ver_app_node),
+    ];
+    block_on({
+        async move {
+            async fn install_via_cr(
+                cr: &UniversalCanister<'_>,
+                target: Principal,
+            ) -> Result<Vec<u8>, AgentError> {
+                cr.update(
+                    wasm().call(
+                        management::install_code(target, UNIVERSAL_CANISTER_WASM)
+                            .with_mode(management::InstallMode::Reinstall),
+                    ),
+                )
+                .await
+            }
+
+            for (cr_node, ce_node) in combinations {
+                let controller_agent = cr_node.build_default_agent_async().await;
+                let controllee_agent = ce_node.build_default_agent_async().await;
+
+                let controller = UniversalCanister::new_with_retries(
+                    &controller_agent,
+                    cr_node.effective_canister_id(),
+                    &logger,
+                )
+                .await;
+                let controllee = UniversalCanister::new_with_retries(
+                    &controllee_agent,
+                    ce_node.effective_canister_id(),
+                    &logger,
+                )
+                .await;
+
+                // before setting the controller, this call must fail.
+                assert_reject(
+                    install_via_cr(&controller, controllee.canister_id()).await,
+                    RejectCode::CanisterReject,
+                );
+
+                set_controller(
+                    &controllee.canister_id(),
+                    &controller.canister_id(),
+                    &controllee_agent,
+                )
+                .await;
+
+                let mgr = ManagementCanister::create(&controllee_agent);
+                assert_http_submit_fails(
+                    mgr.stop_canister(&controllee.canister_id()).call().await,
+                    RejectCode::CanisterError,
+                );
+
+                install_via_cr(&controller, controllee.canister_id())
+                    .await
+                    .unwrap();
+
+                controller
+                    .update(wasm().call(management::stop_canister(controllee.canister_id())))
+                    .await
+                    .unwrap();
+
+                controller
+                    .update(wasm().call(management::delete_canister(controllee.canister_id())))
+                    .await
+                    .unwrap();
+            }
+        }
+    })
 }
