+++ rs/tests/execution/general_execution_test.rs
@@ -1,30 +1,29 @@
-#[rustfmt::skip]
+mod general_execution_tests;
 
 use anyhow::Result;
-
+use execution_system_tests_common::config_system_verified_application_subnets;
+use general_execution_tests::api_tests::node_metrics_history_another_subnet_succeeds;
+use general_execution_tests::api_tests::node_metrics_history_ingress_query_fails;
+use general_execution_tests::api_tests::node_metrics_history_ingress_update_fails;
+use general_execution_tests::api_tests::node_metrics_history_non_existing_subnet_fails;
+use general_execution_tests::api_tests::node_metrics_history_query_fails;
+use general_execution_tests::api_tests::node_metrics_history_update_succeeds;
+use general_execution_tests::api_tests::test_controller;
+use general_execution_tests::api_tests::test_cycles_burn;
+use general_execution_tests::api_tests::test_in_replicated_execution;
+use general_execution_tests::api_tests::test_raw_rand_api;
+use general_execution_tests::big_stable_memory::*;
+use general_execution_tests::canister_heartbeat::*;
+use general_execution_tests::canister_lifecycle::*;
+use general_execution_tests::cycles_transfer::*;
+use general_execution_tests::ingress_rate_limiting::*;
+use general_execution_tests::malicious_input::malicious_input_test;
+use general_execution_tests::nns_shielding::*;
+use general_execution_tests::queries::query_reply_sizes;
+use general_execution_tests::wasm_chunk_store::*;
 use ic_system_test_driver::driver::group::SystemTestGroup;
 use ic_system_test_driver::driver::group::SystemTestSubGroup;
 use ic_system_test_driver::systest;
-use ic_tests::execution::api_tests::node_metrics_history_another_subnet_succeeds;
-use ic_tests::execution::api_tests::node_metrics_history_ingress_query_fails;
-use ic_tests::execution::api_tests::node_metrics_history_ingress_update_fails;
-use ic_tests::execution::api_tests::node_metrics_history_non_existing_subnet_fails;
-use ic_tests::execution::api_tests::node_metrics_history_query_fails;
-use ic_tests::execution::api_tests::node_metrics_history_update_succeeds;
-use ic_tests::execution::api_tests::test_controller;
-use ic_tests::execution::api_tests::test_cycles_burn;
-use ic_tests::execution::api_tests::test_in_replicated_execution;
-use ic_tests::execution::api_tests::test_raw_rand_api;
-use ic_tests::execution::big_stable_memory::*;
-use ic_tests::execution::canister_heartbeat::*;
-use ic_tests::execution::canister_lifecycle::*;
-use ic_tests::execution::config_system_verified_application_subnets;
-use ic_tests::execution::cycles_transfer::*;
-use ic_tests::execution::ingress_rate_limiting::*;
-use ic_tests::execution::malicious_input::malicious_input_test;
-use ic_tests::execution::nns_shielding::*;
-use ic_tests::execution::queries::query_reply_sizes;
-use ic_tests::execution::wasm_chunk_store::*;
 
 fn main() -> Result<()> {
     SystemTestGroup::new()
@@ -53,9 +52,11 @@ fn main() -> Result<()> {
                 .add_test(systest!(create_canister_with_no_controllers))
                 .add_test(systest!(create_canister_with_multiple_controllers))
                 .add_test(systest!(create_canister_with_too_many_controllers_fails))
+                .add_test(systest!(create_canister_with_no_settings_field))
                 .add_test(systest!(create_canister_with_none_settings_field))
                 .add_test(systest!(create_canister_with_empty_settings))
                 .add_test(systest!(create_canister_with_settings))
+                .add_test(systest!(provisional_create_canister_with_no_settings))
                 .add_test(systest!(create_canister_with_freezing_threshold))
                 .add_test(systest!(
                     create_canister_with_invalid_freezing_threshold_fails
