+++ rs/tests/execution/general_execution_tests/big_stable_memory.rs
