+++ rs/tests/execution/general_execution_tests/canister_heartbeat.rs
