+++ rs/tests/execution/general_execution_tests/malicious_input.rs
