+++ rs/tests/execution/general_execution_tests/mod.rs
@@ -1,0 +1,10 @@
+pub mod api_tests;
+pub mod big_stable_memory;
+pub mod canister_heartbeat;
+pub mod canister_lifecycle;
+pub mod cycles_transfer;
+pub mod ingress_rate_limiting;
+pub mod malicious_input;
+pub mod nns_shielding;
+pub mod queries;
+pub mod wasm_chunk_store;
