+++ rs/tests/execution/general_execution_tests/nns_shielding.rs
@@ -9,8 +9,6 @@ use ic_agent::{
 use ic_base_types::RegistryVersion;
 use ic_management_canister_types::SetupInitialDKGArgs;
 use ic_nns_constants::CYCLES_MINTING_CANISTER_ID;
-use ic_registry_subnet_type::SubnetType;
-use ic_system_test_driver::driver::ic::{InternetComputer, Subnet};
 use ic_system_test_driver::driver::test_env::TestEnv;
 use ic_system_test_driver::driver::test_env_api::{GetFirstHealthyNodeSnapshot, HasPublicApiUrl};
 use ic_system_test_driver::{types::*, util::CYCLES_LIMIT_PER_CANISTER, util::*};
@@ -204,12 +202,6 @@ pub fn no_cycle_balance_limit_on_nns_subnet(env: TestEnv) {
             CYCLES_LIMIT_PER_CANISTER * 2u64
         );
     });
-}
-
-pub fn config() -> InternetComputer {
-    InternetComputer::new()
-        .add_subnet(Subnet::fast_single_node(SubnetType::System))
-        .add_subnet(Subnet::fast_single_node(SubnetType::Application))
 }
 
 /// Tests whether creating a canister on a subnet other than self fails when not
